api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pandas
numpy
scikit-learn
sentence-transformers
google-generativeai
